Reorganize Activities/Adapters/Models
