+++ app/src/main/java/com/cs3773/roadrunnergrocery/Activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.cs3773.roadrunnergrocery;
+package com.cs3773.roadrunnergrocery.Activities;
 
 import androidx.appcompat.app.AppCompatActivity;
