+++ app/src/main/java/com/cs3773/roadrunnergrocery/Activities/SearchActivity.java
@@ -1,11 +1,13 @@
-package com.cs3773.roadrunnergrocery;
+package com.cs3773.roadrunnergrocery.Activities;
 
 import android.os.Bundle;
+
+import com.cs3773.roadrunnergrocery.R;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-public class CheckoutActivity extends AppCompatActivity {
+public class SearchActivity extends AppCompatActivity {
 
     // Variables here
 
@@ -13,7 +15,7 @@ public class CheckoutActivity extends AppCompatActivity {
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_checkout);
+        setContentView(R.layout.activity_search);
     }
 
     //Other methods here
