+++ app/src/main/java/com/cs3773/roadrunnergrocery/Models/AccountModel.java
@@ -1,0 +1,11 @@
+package com.cs3773.roadrunnergrocery.Models;
+
+public class AccountModel {
+
+    // Class Variables here
+
+    // Constructor(s) here
+
+    // Methods here
+
+}
