+++ app/src/main/java/com/cs3773/roadrunnergrocery/Models/CardModel.java
@@ -1,0 +1,11 @@
+package com.cs3773.roadrunnergrocery.Models;
+
+public class CardModel {
+
+    // Class Variables here
+
+    // Constructor(s) here
+
+    // Methods here
+
+}
