+++ app/src/main/java/com/cs3773/roadrunnergrocery/Models/ItemModel.java
@@ -1,0 +1,11 @@
+package com.cs3773.roadrunnergrocery.Models;
+
+public class ItemModel {
+
+    // Class Variables here
+
+    // Constructor(s) here
+
+    // Methods here
+
+}
